Test Grid getListDataView
Test that regetting data view
after provider changed to wrong
type throws exception.
